Tic Tac Toe Human v. Human
This is the first version of a tic tac toe game is Ruby. It is a 3X3
board and meant to be played in the console. It allows the users to
select their symbols, dsiplays the board, and select their move. It also
communicates if the game was won or tied